requirements.txt
Django==1.10
django-filebrowser==3.7.2
django-graphiql==0.4.4
django-grappelli==2.8.1
django-summernote==0.8.5
graphene==0.10.2
graphql-core==0.5.3
graphql-django-view==1.3
graphql-relay==0.4.4
iso8601==0.1.11
Pillow==3.3.0
promise==0.4.2
singledispatch==3.4.0.3
six==1.10.0
